refactor: make small changes

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -17,27 +17,26 @@ public class DirectorService {
     private final DirectorDbStorage directorDbStorage;
 
     public List<Director> getAllDirectors() {
+        log.info("Получение списка режиссеров.");
         return directorDbStorage.getAllDirectors();
     }
 
     public Director getDirectorById(int id) {
-        Director director = directorDbStorage.getDirectorById(id);
-        if (director != null) {
-            log.info("Найден режиссер: {} {}", director.getId(), director.getName());
-        } else {
-            log.info("Режиссер с id {} не найден.", id);
+        if (!isDirectorPresent(id)) {
             throw new DataNotFoundException(String.format("Режиссер с id: %d не найден.", id));
         }
-        return director;
+        log.info("Получение режиссера с id: {} .", id);
+        return directorDbStorage.getDirectorById(id);
     }
 
     public Director addDirector(Director director) {
+        log.info("Добавление режиссера : {} .", director.getName());
         return directorDbStorage.addDirector(director);
     }
 
     public Director updateDirector(Director director) {
         Director updatedDirector;
-        if (directorDbStorage.getDirectorById(director.getId()) == null) {
+        if (!isDirectorPresent(director.getId())) {
             log.warn("Попытка обновить несуществующего режиссера с id: {}", director.getId());
             throw new DataNotFoundException(String.format("Режиссер с id: %d не найден.", director.getId()));
         }
@@ -46,7 +45,21 @@ public class DirectorService {
         return updatedDirector;
     }
 
-    public boolean deleteDirector(int id) {
-        return directorDbStorage.deleteDirector(id);
+    public void deleteDirector(int id) {
+        if (!isDirectorPresent(id)) {
+            throw new DataNotFoundException(String.format("Режиссер с id: %d не найден.", id));
+        }
+        log.info("Удаление режиссера с id: {} .", id);
+        directorDbStorage.deleteDirector(id);
+    }
+
+    public boolean isDirectorPresent(int id) {
+        boolean isPresent = directorDbStorage.existsById(id);
+        if (isPresent) {
+            log.info("Найден режиссер с id: {}", id);
+        } else {
+            log.info("Режиссер с id {} не найден.", id);
+        }
+        return isPresent;
     }
 }
